refactor: remove unused properties and streamline component states
