+++ packages/components/src/internal/functional-components/progress/api.tsx
@@ -1,14 +1,12 @@
-import type { LabelProp, MaxProp, UnitProp, NumberValueProp, VariantProgressProp } from '../../props';
-import type { ComponentApi, InternalOf } from '../generic-types';
+import type { LabelProp, MaxProp, NumberValueProp, UnitProp, VariantProgressProp } from '../../props';
+import type { ComponentApi } from '../generic-types';
 
 export interface ProgressApi extends ComponentApi {
 	Props: {
 		Optional: LabelProp & UnitProp & VariantProgressProp;
 		Required: MaxProp & NumberValueProp;
 	};
-	States: InternalOf<UnitProp> &
-		InternalOf<VariantProgressProp> & {
-			liveValue: number;
-			max: number;
-		};
+	States: {
+		liveValue: number;
+	};
 }
