+++ packages/components/src/internal/props/index.ts
@@ -3,7 +3,6 @@ export * from './color';
 export * from './count';
 export * from './helpers/factory';
 export * from './href';
-export * from './initials';
 export * from './label';
 export * from './loading';
 export * from './max';
